PoW methods: remove txsWithPoW parameter
In the IotaClient::attachToTangle() and PoWClient::pow() methods,
this parameter was the list to be filled by transactions with PoW
data. In order to avoid keeping in memory both the transactions
without PoW and the transactios with PoW, we now use the txs
argument to hold the list of transactions with PoW (which overwrite
the transactions without PoW that were passed when calling the
methods).