IotaClient: add JsonHttpClient helper class in all platforms
The ArduinoJson serializer sends data to the output stream one character at a time; depending on the network client in use, this may lead to an inefficent fragmentation of HTTP requests into multiple TCP packets. In order to avoid this issue, the JsonHttpClient helper class is now implemented in all platforms (not just in the ESP8266 platform), so that the output data is buffered in chunks optimally sized for TCP packets.