IotaWallet transaction bundle: avoid double allocation of transaction data
Instead of allocating bundle transaction data in the iotaWalletBundle structure and then duplicating it in the std::vector of Strings to be passed to IotaClient::attachToTangle(), transaction data is put directly in the vector of Strings, thus avoiding double allocation and reducing dynamic memory requirements. The iotaWalletTxReceiver static function achieves this by calling String::reserve() to allocate in a String object the memory needed for each transaction.