iotaNodeInfo: update struct members
Now that non-Java implementations of IOTA nodes are out in the
wild, a node response to a getNodeInfo request may not contain
JRE-related attributes; also, the HORNET node software does not
return a packetsQueueSize attribue either.
This commit removes from the iotaNodeInfo structure members that
may not be present in a getNodeInfo response, and adds a couple of
new members (milestoneStartIndex and lastSnapshottedMilestoneIndex)
that are now present in modern node implementations.